Two branches were started from the same pokeheartgold commit and never saw
each other. This one worked down from konefr's playable content; the other
worked up from what hg-engine's hooks land on, and converted 21 functions
from assembly to C on the way. They touched six files in common and
disagreed in one place, so the merge is worth making now rather than later.

What the other branch brings that this one wanted:

  - the ability field widened past one byte through the save, the battle
    record, the AI's memory, the PC box, the summary, the Frontier records,
    the Pokewalker and the Trainer House. The comment in abilities.h
    numbering the added abilities from 124 because "the saved record still
    keeps an ability in one byte" was describing exactly this constraint.
  - the three ability message banks, 960 texts
  - reusable repels, one of the behaviours this branch had still open
  - the Rage and Fire Fang / Shadow Force fixes
  - the hook inventory and its ledger

Two behaviours had been done on both sides. The friendship threshold agreed.
Overworld poison did not: this branch stopped it at the caller, the other
emptied ApplyPoisonStep itself. The other is kept, because it is what the
reference does — a one-byte patch setting the status mask to zero inside that
function, so the game still walks the party every fourth step and simply
finds nothing poisoned. FieldSystem_UpdatePoison goes back to what upstream
wrote.

The one conflict was GetBattlerAbility, widened to u16 on one side and given
the Neutralizing Gas check on the other; it keeps both.

Three of the merged tests compile source out of the tree and so had to be
told about this branch's additions: the ability-width test now extracts the
Neutralizing Gas helper, and the friendship test stubs MonHasMoveOfType.

Both ROMs build with no new warnings and all 103 tests pass.

+++ include/battle/overlay_12_0224E4FC.h
@@ -15,6 +15,7 @@ void ov12_0224ED00(BattleContext *ctx, int id, int battlerId, int index);
 BOOL Link_QueueNotEmpty(BattleContext *ctx);
 void BattleBuffer_Clear(BattleContext *ctx, int battlerId);
 int GetBattlerVar(BattleContext *ctx, int battlerId, u32 varId, void *data);
+// BMON_DATA_ABILITY consumes a u16 payload.
 void SetBattlerVar(BattleContext *ctx, int battlerId, u32 varId, void *data);
 void AddBattlerVar(BattleContext *ctx, int battlerId, u32 varId, int data);
 void BattleMon_AddVar(BattleMon *mon, u32 varId, int data);
@@ -54,7 +55,7 @@ BOOL CurseUserIsGhost(BattleContext *ctx, u16 moveNo, int battlerId);
 BOOL CanStealHeldItem(BattleSystem *battleSystem, BattleContext *ctx, int battlerId);
 BOOL CanTrickHeldItem(BattleContext *ctx, int battlerId);
 BOOL WhirlwindCheck(BattleSystem *battleSystem, BattleContext *ctx);
-u8 GetBattlerAbility(BattleContext *ctx, int battlerId);
+u16 GetBattlerAbility(BattleContext *ctx, int battlerId);
 BOOL CheckBattlerAbilityIfNotIgnored(BattleContext *ctx, int battlerIdAttacker, int battlerIdTarget, int ability);
 BOOL CanSwitchMon(BattleSystem *battleSystem, BattleContext *ctx, int battlerId);
 BOOL CantEscape(BattleSystem *battleSystem, BattleContext *ctx, int battlerId, BattleMessage *msg);
